Moving some files to the proper folders
